Refactor external link component name
