+++ src/components/Organisms/Footer/Footer.tsx
@@ -1,5 +1,5 @@
 import { Footer } from './Footer.styles.ts';
-import Link from '../../Atoms/Link/Link.tsx';
+import ExtLink from '../../Atoms/ExtLink/ExtLink.tsx';
 import Copyright from '../../Molecules/Copyright/Copyright.tsx';
 import { data } from '../../../utils/data.ts';
 
@@ -7,7 +7,7 @@ const FooterWrapper = () => {
   return (
     <Footer>
       <p>RCM4Strapi</p>
-      <Link url={data.blogUrl} name={data.blogName} size="s" />
+      <ExtLink url={data.blogUrl} name={data.blogName} size="s" />
       <Copyright />
     </Footer>
   );
